Edit Genre

+++ app/Http/Controllers/GenreController.php
@@ -3,11 +3,34 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Str;
 use App\Genre;
 use App\AnimeGenre;
 use App\Anime;
+use Illuminate\Support\Facades\Auth;
 class GenreController extends Controller
 {
+    public function __construct(){
+      $this->middleware('auth:api',['only'=>[
+        'store',
+        'update',
+        'delete',
+        'indexDelete',
+        'harddestroy',
+        'destroy',
+        'restore',
+      ]]);
+      $this->middleware('role',['only'=>[
+        'store',
+        'update',
+        'delete',
+        'indexDelete',
+        'harddestroy',
+        'destroy',
+        'restore',
+      ]]);
+    }
     /**
      * Display all listing of the resource.
      *
@@ -15,10 +38,10 @@ class GenreController extends Controller
      */
     public function index()
     {
-      $genre=Genre::select('id','genre')->get();
+      $genre=Genre::select('id','genre','slug')->get();
       foreach($genre as $i){
         $i->view_genre=[
-          'href'=>'/api/v1/genre/'.$i->id,
+          'href'=>'/api/v1/genre/'.$i->slug,
           'method'=>'GET',
         ];
       }
@@ -37,11 +60,11 @@ class GenreController extends Controller
      */
     public function indexDelete()
     {
-      $genre= Genre::onlyTrashed()->select('id','genre')->get();
+      $genre= Genre::onlyTrashed()->select('id','genre','slug')->get();
       if($genre!=null){
         foreach($genre as $i){
           $i->view_genre=[
-            'href'=>'/api/v1/genre/'.$i->id.'/delete',
+            'href'=>'/api/v1/genre/'.$i->slug.'/delete',
             'method'=>'GET',
           ];
         }
@@ -66,11 +89,12 @@ class GenreController extends Controller
     public function store(Request $request)
     {
         $genre= new Genre([
-          'genre'=>$request->genre
+          'genre'=>$request->genre,
+          'slug'=>Str::slug($request->genre),
         ]);
         if($genre->save()){
           $genre->view_genre=[
-            'href'=>'/api/v1/genre/'.$genre->id,
+            'href'=>'/api/v1/genre/'.$genre->slug,
             'method'=>'GET',
           ];
           $response=[
@@ -91,44 +115,49 @@ class GenreController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show($slug)
     {
-      $genre=Genre::select('id','genre')->where('id',$id)->first();
-      $pivot_anime=AnimeGenre::select('anime_id')->where('genre_id',$id)->get();
-      //return isset($pivot_anime);
-      if(count($pivot_anime)>0){
-        foreach($pivot_anime as $i){
-          $anime[]=Anime::select('id','judul','status','cover')->where('id',$i->anime_id)->first();
+      $genre=Genre::select('id','genre','slug')->where('slug',$slug)->first();
+      if($genre!=null){
+        $pivot_anime=AnimeGenre::select('anime_id')->where('genre_id',$genre->id)->orderBy('created_at','desc')->get();
+        if(count($pivot_anime)>0){
+          foreach($pivot_anime as $i){
+            $anime[]=Anime::select('id','judul','slug','cover')->where('id',$i->anime_id)->first();
+          }
+          foreach($anime as $i){
+            $i->cover='/api/v1/anime/'.$i->slug.'/cover';
+          }
         }
-        foreach($anime as $i){
-          $i->cover='/api/v1/anime/'.$i->id.'/cover';
+        if(count($pivot_anime)>0){
+          $response=[
+            'message'=>'Detail Genre',
+            'data'=>$genre,
+            'anime'=>$anime,
+          ];
+        }else{
+          $response=[
+            'message'=>'Detail Genre',
+            'data'=>$genre,
+            'anime'=>null,
+          ];
         }
+        return response()->json($response,200);
       }
-      if(count($pivot_anime)>0){
-        $response=[
-          'message'=>'Detail Genre',
-          'data'=>$genre,
-          'anime'=>$anime,
-        ];
-      }else{
-        $response=[
-          'message'=>'Detail Genre',
-          'data'=>$genre,
-          'anime'=>null,
-        ];
-      }
-      return response()->json($response,200);
+      $response=[
+        'message'=>'Genre '.$slug.' Tidak ditemukan !'
+      ];
+      return response()->json($response,404);
     }
 
     /**
     *Show Detail Delete with soft delete
     */
-    public function showDelete($id)
+    public function showDelete($slug)
     {
-      $genre= Genre::onlyTrashed()->select('id','genre')->where('id',$id)->first();
+      $genre= Genre::onlyTrashed()->select('id','genre','slug')->where('slug',$slug)->first();
       if($genre!=null){
         $genre->restore=[
-          'href'=>'/api/v1/genre/'.$genre->id.'/restore',
+          'href'=>'/api/v1/genre/'.$genre->slug.'/restore',
           'method'=>'GET',
         ];
         $response=[
@@ -138,7 +167,7 @@ class GenreController extends Controller
         return response()->json($response,201);
       }
       $response=[
-        'message'=>'Telah terjadi kesalahan'
+        'message'=>'Genre '.$slug.' Tidak ditemukan !'
       ];
       return response()->json($response,404);
     }
@@ -149,23 +178,30 @@ class GenreController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, $slug)
     {
-        $genre= Genre::select('id','genre')->where('id',$id)->first();
-        $genre->genre=$request->genre;
-        if($genre->update()){
-          $genre->view_genre=[
-            'href'=>'/api/v1/genre/'.$genre->id,
-            'method'=>'GET',
-          ];
+        $genre= Genre::select('id','genre')->where('slug',$slug)->first();
+        if($genre!=null){
+          $genre->genre=$request->genre;
+          $genre->slug=Str::slug($request->genre);
+          if($genre->update()){
+            $genre->view_genre=[
+              'href'=>'/api/v1/genre/'.$genre->id,
+              'method'=>'GET',
+            ];
+            $response=[
+              'message'=>'Genre Berhasil di ubah ke '.$genre->genre,
+              'data'=>$genre
+            ];
+            return response()->json($response,201);
+          }
           $response=[
-            'message'=>'Genre Berhasil di ubah ke '.$genre->genre,
-            'data'=>$genre
+            'message'=>'Telah terjadi kesalahan'
           ];
-          return response()->json($response,201);
+          return response()->json($response,404);
         }
         $response=[
-          'message'=>'Telah terjadi kesalahan'
+          'message'=>'Genre '.$slug.' Tidak ditemukan !'
         ];
         return response()->json($response,404);
     }
@@ -176,22 +212,28 @@ class GenreController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function restore($id)
+    public function restore($slug)
     {
-      $genre= Genre::onlyTrashed()->select('id','genre')->where('id',$id)->first();
-      if($genre->restore()){
-        $genre->view_genre=[
-          'href'=>'/api/v1/genre/'.$genre->id,
-          'method'=>'GET',
-        ];
+      $genre= Genre::onlyTrashed()->select('id','genre','slug')->where('slug',$slug)->first();
+      if($genre!=null){
+        if($genre->restore()){
+          $genre->view_genre=[
+            'href'=>'/api/v1/genre/'.$genre->slug,
+            'method'=>'GET',
+          ];
+          $response=[
+            'message'=>'Genre '.$genre->genre.' Berhasil di Restore',
+            'data'=>$genre
+          ];
+          return response()->json($response,201);
+        }
         $response=[
-          'message'=>'Genre '.$genre->genre.' Berhasil di Restore',
-          'data'=>$genre
+          'message'=>'Telah terjadi kesalahan'
         ];
-        return response()->json($response,201);
+        return response()->json($response,404);
       }
       $response=[
-        'message'=>'Telah terjadi kesalahan'
+        'message'=>'Genre '.$slug.' Tidak ditemukan !'
       ];
       return response()->json($response,404);
     }
@@ -202,22 +244,29 @@ class GenreController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($slug)
     {
-        $genre= Genre::select('id','genre')->where('id',$id)->first();
-        if($genre->delete()){
-          $genre->restore=[
-            'href'=>'/api/v1/genre/'.$genre->id.'/restore',
-            'method'=>'GET',
-          ];
+        $genre= Genre::select('id','genre','slug')->where('slug',$slug)->first();
+        if($genre!=null){
+          if($genre->delete()){
+            $genre->restore=[
+              'href'=>'/api/v1/genre/'.$genre->slug.'/restore',
+              'method'=>'GET',
+            ];
+            $response=[
+              'message'=>'Genre '.$genre->genre.' Berhasil di Hapus',
+              'data'=>$genre
+            ];
+            return response()->json($response,201);
+          }
           $response=[
-            'message'=>'Genre '.$genre->genre.' Berhasil di Hapus',
-            'data'=>$genre
+            'message'=>'Telah terjadi kesalahan'
           ];
-          return response()->json($response,201);
+          return response()->json($response,404);
         }
+
         $response=[
-          'message'=>'Telah terjadi kesalahan'
+          'message'=>'Genre '.$slug.' Tidak ditemukan !'
         ];
         return response()->json($response,404);
     }
@@ -228,25 +277,31 @@ class GenreController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function harddestroy($id)
+    public function harddestroy($slug)
     {
-      $genre= Genre::select('id','genre')->where('id',$id)->first();
-      if($genre->forceDelete()){
-        $genre->create=[
-          'href'=>'/api/v1/genre/',
-          'param'=>[
-            'genre'=>'text'
-          ],
-          'method'=>'POST',
-        ];
+      $genre= Genre::onlyTrashed()->select('id','genre')->where('slug',$slug)->first();
+      if($genre!=null){
+        if($genre->forceDelete()){
+          $genre->create=[
+            'href'=>'/api/v1/genre/',
+            'param'=>[
+              'genre'=>'text'
+            ],
+            'method'=>'POST',
+          ];
+          $response=[
+            'message'=>'Genre '.$genre->genre.' Berhasil di Hapus Secara Permanent',
+            'data'=>$genre
+          ];
+          return response()->json($response,201);
+        }
         $response=[
-          'message'=>'Genre '.$genre->genre.' Berhasil di Hapus Secara Permanent',
-          'data'=>$genre
+          'message'=>'Telah terjadi kesalahan'
         ];
-        return response()->json($response,201);
+        return response()->json($response,404);
       }
       $response=[
-        'message'=>'Telah terjadi kesalahan'
+        'message'=>'Genre '.$slug.' Tidak ditemukan !'
       ];
       return response()->json($response,404);
     }
